requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
mysql-connector-python==8.2.0
Pillow==10.2.0

